python_scraper_api/pyproject.toml
[project]
name = "walmart-initial-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "selenium>=4.33.0",
]
